requirements.txt
alembic==1.10.4
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
colorama==0.4.6
Flask==2.3.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask-swagger-ui==4.11.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
passlib==1.7.4
PyJWT==2.6.0
pytz==2023.3
six==1.16.0
SQLAlchemy==2.0.12
typing_extensions==4.5.0
Werkzeug==2.3.3
